Version 1.1.2 Update
In the OptionsView the path for the Fourier series can be drawn with an alpha gradient, creating the effect of a fading trail by specifying the length of the trail. Access the OptionsView view by tapping on the sliders button in the toolbar of the Graphics view. Then select the tab view labelled Σ, and set the Trail Length. A length of 0 turns off the effect. The effect is demonstrated in the new animated GIF that appears in the startup view.